backend/requirements.txt
fastapi
uvicorn
PyGithub
gitpython
pipreqs
pdoc
flake8
radon
pylint
python-multipart 